正式赛benchmark探测

+++ src/main/java/io/openmessaging/ValueReader.java
@@ -65,7 +65,7 @@ public class ValueReader {
             buffer.clear();
         }
         long t = message.getA() - message.getT();
-        if (t >= Short.MAX_VALUE || t <= Short.MIN_VALUE) {
+        if (t >= Integer.MAX_VALUE || t <= Integer.MIN_VALUE) {
             System.out.println("more than");
         }
         buffer.putInt((int) (message.getA() - message.getT()));
